third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/029-expected.txt

This is a testharness.js-based test.
FAIL  scheduler: javascript: URL in HREF assert_array_equals: property 1, expected "end script #1" but got "JS URL"\n\nassert_array_equals: property 1, expected "end script #1" but got "JS URL"
Harness: the test ran to completion.
